fix(C#): 修复 Kbdllhookstruct 字段类型错位导致 CapsLock 不被拦截
根因：KBDLLHOOKSTRUCT 的 vkCode/scanCode 是 DWORD(4字节)，但 C# 声明成
ushort(2字节)，导致 Vk 之后的 Scan/Flags 全部错位——kb.Flags 实际读到的是
scanCode 字段，kb.Scan 读的是 vkCode 高位字节(恒0)。

后果（用户报告：按下 CapsLock 立即开启大写键、所有 CapsLock+ 失效）：
注入检查 (kb.Flags & LLKHF_INJECTED) 读的是 scanCode。物理 CapsLock 扫描码
0x3A，0x3A & 0x10 != 0 → 误判为注入事件 → 直接放行不吞 → 灯翻转。
同理 E(0x12)/W(0x11)/R(0x13) 等扫描码 bit0x10 置位的键，CapsLock+ 组合全被
误判注入放行 → 失效。kb.Scan 恒0 也使帮助面板 SC029 派发一直失效。

修复：Vk/Scan 改 uint(DWORD) 对齐 Win32 布局；回调引入局部 ushort vk 给下游
ushort 签名方法。

验证：asInvoker 注入 CapsLock(扫描码0x3A) 修复前 flags 读成 58(=0x3A)且被
误判注入放行；修复后 keydown flags=16(0x10 LLKHF_INJECTED)、keyup flags=144
(0x10|0x80 LLKHF_UP)，注入检查正确触发。物理键 flags=0 → 不再误判 → 进入
状态机被吞。此 bug 自阶段1 起存在，阶段1-4 所有 CapsLock+ 功能一并修复。

诊断用 UI 自动化 + keybd_event 注入定位（钩子按设计忽略 SendInput 的注入标志，
但 keybd_event 在该机器上 flags 读 0，扫描码 0x3A 复现了物理键的错位行为）。

+++ csharp/Native/Win32.cs
@@ -150,11 +150,11 @@ internal static class Win32
     [StructLayout(LayoutKind.Sequential)]
     public struct Kbdllhookstruct
     {
-        public ushort Vk;
-        public ushort Scan;
-        public uint Flags;
-        public uint Time;
-        public IntPtr ExtraInfo;
+        public uint Vk;        // vkCode，偏移 0（DWORD）
+        public uint Scan;      // scanCode，偏移 4（DWORD）
+        public uint Flags;     // flags，偏移 8（DWORD）
+        public uint Time;      // 偏移 12
+        public IntPtr ExtraInfo; // 偏移 16（x64）
     }
 
     /// <summary>低级鼠标钩子事件数据（WH_MOUSE_LL）。</summary>
